# Add performance logger and use to keep track of processing a block