Adds a new option `perf` which is a number to set the amount of performance logging. Default is nothing

Example output:
```
Block 501: 6979 ms
   getBlockHeight(): 1012 ms
     getBlockHash(): 200 ms
     getBlockByHash(): 812 ms
   events/specVersion: 1356 ms
   wrapBlock: 1 ms
   in queue: 3696 ms
   indexBlock(): 445 ms
     getPatchedApi: 6 ms
     indexBlockData(): 425 ms
       block: 0 ms
       initialization events: 0 ms
       extrinsics and events: 424 ms
       finalization events: 0 ms
     poi/mmr: 0 ms
     commitTx: 6 ms
```