@ianhe8x 2022-05-25
I always wanted to find a lib in node that can generate Histogram.