@stwiname 2022-05-25
> I always wanted to find a lib in node that can generate Histogram.

I've been using https://clinicjs.org/documentation/flame/03-first-analysis/ if thats what you're after